OLH-874: Give the AM API stub a domain
The firewall rules on the frontend app stop it from making requests to
any non-`*.gov.uk` domains.

Give the account management API stub a domain in our dev and build
environments so the frontend can reach it.

This is following the same pattern as we used for the domain at the
frontend (see https://github.com/alphagov/di-account-management-frontend/blob/main/deploy/template.yaml#L736-L770).