OLH-1027: Use a real page for logout
Currently when a user clicks the 'Sign out' button using the stub they
end up on an error page because the OIDC end session endpoint in our
configuration file doesn't exist.

Rather than writing a new lambda as part of this stub API, send the user
directly to the Auth team's frontend to see the sign out page. This
intentionally uses the build environment as there isn't an equivalent
for dev.