OLH-1026: Fix phone_number claim names
The frontend is currently not storing the user's phone number in the
session once it's made the request to the `/userinfo` endpoint.

This means we're not displaying the 'change your phone number' flow and
the frontend is behaving as if the user has an authenticator app.

Our [public documentation][1] says that the phone number claim is called
`phone`, however the [OIDC specification][2] says it should be `phone_number`.

The same is true for the `phone_number_verified` claim.

This is a mistake in our documentation. Update the claim name to match what's
actually set.

[1]: https://docs.sign-in.service.gov.uk/integrate-with-integration-environment/integrate-with-code-flow/#retrieve-user-information
[2]: https://openid.net/specs/openid-connect-core-1_0.html#StandardClaims