OLH-1062 - Adjusting reserved concurrency to support performance testing.
We have a requirement to support 100 requests per second to the delete account function which uses various
functions in the stubs.  Assuming each function is called once for each request and it takes approx 300ms
to complete then the reserved concurrency should be 100 * 0.3 = 30.  This is a crude first estimate that
is intended to get to the point where the performance test runs without the current 95% failure rate.