OLH-1619: Restructure Lambda files
Restructure this repo so it's a single NPM package with separate files
for each Lambda function.

There's two different services that are stubbed from this repo - the
account management API and the OIDC API. I've put these (and their tests)
into separate folders so it's easier to see which handler belongs to
which service.