BAU: Remove `updated_at` from userinfo response
This field is an optional one in the OIDC spec. Our public tech docs say
it's present in our real response, but this is a mistake. In the real
version of our OIDC IdP we don't include `updated_at`.

I couldn't find any reference to us using this value in the frontend
(which makes sense if it's not provided) so this PR won't break our
pre-staging environments.