BAU: Consolidate lambda roles and policies (#205)
## Proposed changes

### What changed
Many of our stub lambdas have the same permissions requirements; we have
lots with no extra permissions and all the method management ones need
permission to query the DynamoDB table.

Reduce the complexity and length of our template by:
- Adding a base execution role with no extra AWS permissions and using
this where possible
- Adding a single managed policy that grants permission to read from the
DynamoDB table
- Adding a single execution role that has the base permissions and the
policy to read from the DynamoDB table

This allows us to delete a lot of the duplicate per-lambda roles and IAM
policies. The permissions for each lambda hasn't changed as a result of
this; it's purely how we organise and assign those permissions.

### Why did it change

<!-- Describe the reason these changes were made - the "why" -->

### Related links

<!-- List any related PRs -->
<!-- List any related ADRs or RFCs -->

## Checklists

<!-- Merging this PR deploys the stubs. Please answer accurately. -->

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed


### Permissions

- [x] This PR adds or changes permissions

## Testing

I'll deploy this branch to dev and check the stubs still work as
expected.