OLH-2538: Upgrade Lambda runtime to Node 22 (#439)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->

### What changed

From Node 18 as that's about to be deprecated. Our policy[^1] says we
should be on the most recent LTS runtime, so that's Node 22.

This is a safe upgrade as we already run our tests locally and in CI
using Node 22.

I've taken the opportunity to move this parameter to the `Globals`
object as all these Lambdas use the same runtime.

[^1]: https://team-manual.account.gov.uk/policies/#active-policies
### Why did it change

Node 18 is EOL in September