OLH-2534: Update the MFA API url structure
The URL structure of the API has changed to include the public subject
ID in the path, rather than the API inferring it from the auth token.

Update the URLs in our stubs to match and pick up the new path parameter
in the retrieve handler.

We'll also need to update the other methods to use the ID from the path.
My plan is to validate this change with the retrieve method, then I'll
udpate the rest in a followup PR.