OLH-2534: Use a unique user ID for each scenario
We think this is the cause of a caching bug where switching between
scenarios doesn't change the MFA methods in the frontend.

Previously all the scenarios had the same user ID and the frontend
sesssion store has an index on session ID, so we think some parts of the
session are being reused.