OLH-2534: Disable caching on the method management stub
Previously this stub had no GET request routes so the default caching
had no effect. Now the retrieve method has swtiched to a GET we were
seeing odd responses on the frontend. We thought they were due to
session reuse since all scenarios had the same user ID, but actually the
frontend was making a new API call but it was hitting the cache and
getting the wrong response for the scenario