OLH-2655: Handle OIDC logout requests
'Handle' OIDC logout requests that include a `post_logout_redirect_uri`
parameter by validating the domain and simply redirecting the user back
to that URI.

This is compliant with the logout interface described in [our public
tech documentation](https://docs.sign-in.service.gov.uk/integrate-with-integration-environment/managing-your-users-sessions/#make-a-request-to-log-your-user-out-of-gov-uk-one-login).

We need to do the domain validation here to protect ourselves from an
[open redirect attack](https://hacksplaining.com/prevention/open-redirects).
We need to be extra carfeul here as we allow redirects to `localhost`.
If an attacker's got a webserver running on a user's machine then that
user is already having a bad day. We need to do as much as we can to
prevent phishing links off the `*.gov.uk` domain which is why I've also
added the referrer check.