OLH-2566: Allow ID token and redirect URI to be missing
In the real implementation of GOV.UK One Login these parameters can
both me missing. In that case, Orchestration will redirect users to the
generic 'Signed out' page rather than back to the relying party.

Update our stubs to handle the same behaviour. This handler will serve
multiple environments (we use our dev environment stubs when running the
frontend locally), so we need to construct the default return URL. We're
already relying on at least one of the Origin or Referer headers being
present, so use these to build it.