OLH-2655: Don't infer default redirect URI from headers
We decided this was too clever. Instead now build the default URL using
the same logic as the real service.

This does mean that if we logout from localhost we'll get redirected to
the signed out page in Auth's dev environment. That's manageable - the
same sort of thing happens now anyway.